Gestion d'accessibilite des pagesen fonction de l'utilisateur et affichage de l'auteur

+++ delete-article.php
@@ -4,6 +4,14 @@
  * @var ArticleDAO
  */
 require_once("./database/database.php");
+require_once __DIR__.'/database/security.php';
+
+$currentUser = isLoggedIn();
+
+if(!$currentUser) {
+    header('Location: /');
+}
+
 $articleDAO = require_once './database/models/ArticleDAO.php';
 
 
@@ -11,6 +19,10 @@ $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
 $id = $_GET['id'] ?? '';
 
 if($id) {
-    $articleDAO->deleteOne($id);
+    $article = $articleDAO->getOne($id);
+
+    if ($currentUser['id'] === $article['author']) {
+        $articleDAO->deleteOne($id);
+    }
 }
 header('Location: /');
